extract Donation and PhantomFunding
